requirements/base.txt
celery==5.4.0
Django==4.2.16
dependency-injector==4.45.0
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.4.0
ipython==8.32.0
psycopg2-binary==2.9.10
python-json-logger==3.2.1
redis==5.0.1
ruff==0.9.4
sentry-sdk==2.20.0
